Rooms with monsters must have Monster Class
Going forward, rooms with a monster must contain a monster class.
If a room contains treasure, it can't contain a monster or an exit door.

WelcomePlayer() welcomes the player to the room, displaying a custom message in the Console. This lets the player know the name of the room, description of the room and the monster that is in the room