Change equipped weapon
Player is given the option to change the weapon that their character has equipped. They have the choice of any weapon in their inventory. If they choose to change weapon, their current equipped weapon is added to their inventory and the new weapon is removed from inventory and added to the player's equipped weapon.